Add 3-bit pixel art assets for grid rendering
Add all 3-bit pixel art asset aseprite files along with slso8
palette and png spritesheet containing assets required for a
minimal render of a sample grid.

Given intention to print puzzles in black and white/monochrome
format, and given the slightly cluttered appearance of many parts
of the committed assets it is unlikely that they will be used in
production, and will likely be replaced with simpler 1-bit assets.