int! Rewrite state system using vanilla events
WHAT: Rewrite the state and event management using vanilla JS custom
events, removing Boutique as a dependency.

WHY: Boutique does not support Maps or Sets used in impostori puzzle
object, and working with the API as a result was becoming messy.
Additionally, the state management required in this app is simple, and
easily within the scope of the built-in DOM events.